feat(spider): programming language is a percentage

+++ spider/spider/Models/ProgrammingLanguage.cs
@@ -2,13 +2,13 @@
 
 public struct ProgrammingLanguage
 {
-    public ProgrammingLanguage(string name, int size)
+    public ProgrammingLanguage(string name, float size)
     {
         Name = name;
         Size = size;
     }
 
     public string Name { get; }
-    public int Size { get; }
+    public float Size { get; }
    
 }
